refactor(cli): fix issues
